+++ CHANGELOG.md
@@ -27,6 +27,10 @@
 *   Fix a bug where calling `stream.take(0).drain(value)` would not correctly
     forward the `value` through the returned `Future`.
 
+#### `dart:async`
+
+*   Add a `StreamTransformer.fromBind` constructor.
+
 ## 2.1.0-dev.3.0
 
 ### Core library changes